#218 Review fixes: created commons, fixes logical bugs
